+++ packages/vendure-plugin-metrics/src/ui/queries.graphql.ts
@@ -6,6 +6,7 @@ export const GET_METRICS = gql`
       code
       title
       type
+      allowProductSelection
       labels
       series {
         name